pyproject.toml
[project]
name = "e2p2"
version = "0.1.0"
description = "End-to-End PDF parsing"
authors = [
    {name = "Jonas Renault",email = "jonasrenault@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "doclayout-yolo>=0.0.3",
    "huggingface-hub>=0.28.1",
    "numpy>=1.26.4,<2",
    "paddleocr>=2.9.0",
    "paddlepaddle>=2.6.2",
    "pillow>=11.1.0",
    "pypdfium2>=4.30.1",
    "rapid-table>=1.0.3",
    "rapidocr-onnxruntime>=1.4.4",
    "tqdm>=4.67.1",
    "typer>=0.15.1",
    "ultralytics>=8.3.23",
    "unimernet>=0.2.3",
]

[project.scripts]
e2p2 = "e2p2.cli.main:app"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.15.0",
    "pre-commit>=4.1.0",
    "pytest>=8.3.4",
    "ruff>=0.9.6",
]
notebooks = [
    "jupyterlab>=4.3.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line_length = 90
target_version = ['py312']

[tool.ruff]
exclude = [
    ".venv",
    ".git",
    "__pycache__",
    "build",
    "dist",
    "venv",
]
src = ["e2p2", "tests"]
line-length = 90

[tool.ruff.lint]
ignore = []
select = [
    "E",
    "F",
    "W",
    "I001",
]

[tool.ruff.lint.isort]
known-first-party = ["e2p2"]

[tool.mypy]
files = ["e2p2", "tests"]
disable_error_code = ["import-untyped"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::UserWarning",
    'ignore::DeprecationWarning',
]
